fix: I-감사 티어 B 충실도/에딧 7건 수정(#19 의도적 미구현)
stable 코드 대조 기반 충실도 수정 7건:
- #11 SliderCurve: 길이 클램프 trim-only → trim+extend (과길이 슬라이더 볼 타이밍/끝위치)
- #14 CursorRenderer: old-style 트레일 일반 블렌드(가산 아님)
- #15 CursorRenderer: new-style 트레일 회전 미적용(old-style만 CursorTrailRotate)
- #16 CursorRenderer: 트레일 폭<5px 비활성 가드 추가
- #17 HudRenderer: 에러바 히트영역 상단을 화살표 기준으로 보정
- #18 HudRenderer/OsuGlRenderer/OverlayForm: HUD depth 0.5→0.95, 즉시모드 도형 PostDraw로 이동(게임플레이 위)
- #20 SkinManager: ParseColor 범위밖 성분 byte wrap(stable 256 모듈로)

빌드 경고 0/오류 0, 테스트 15/15 통과.

+++ OsuEnlightenOverlay/Skinning/SkinManager.cs
@@ -301,17 +301,21 @@ namespace OsuEnlightenOverlay.Skinning
             string[] split = value.Split(',');
             try
             {
+                // stable Section.ConvertString은 각 성분을 (byte)Convert.ToInt32로 256 모듈로 wrap한다.
+                // int.Parse 후 Color.FromArgb(int,...)는 0~255 밖이면 ArgumentException→catch→Empty로
+                // 색이 통째로 버려졌다. byte 캐스팅으로 stable처럼 wrap(예: 300→44, -5→251)해
+                // 범위 밖 값을 쓰는 손상 스킨도 stable과 동일하게 표시한다 (I-감사 #20).
                 if (split.Length == 3)
                     return Color.FromArgb(
-                        int.Parse(split[0].Trim()),
-                        int.Parse(split[1].Trim()),
-                        int.Parse(split[2].Trim()));
+                        (byte)int.Parse(split[0].Trim()),
+                        (byte)int.Parse(split[1].Trim()),
+                        (byte)int.Parse(split[2].Trim()));
                 if (split.Length == 4)
                     return Color.FromArgb(
-                        allowAlpha ? int.Parse(split[3].Trim()) : 255,
-                        int.Parse(split[0].Trim()),
-                        int.Parse(split[1].Trim()),
-                        int.Parse(split[2].Trim()));
+                        allowAlpha ? (byte)int.Parse(split[3].Trim()) : 255,
+                        (byte)int.Parse(split[0].Trim()),
+                        (byte)int.Parse(split[1].Trim()),
+                        (byte)int.Parse(split[2].Trim()));
             }
             catch { }
             return Color.Empty;
